+++ llvm/lib/Analysis/HashRecognize.cpp
@@ -84,7 +84,9 @@ static bool containsUnreachable(const Loop &L,
   SmallVector<const Instruction *, 16> Worklist(Roots);
   while (!Worklist.empty()) {
     const Instruction *I = Worklist.pop_back_val();
-    Visited.insert(I);
+    // Skip this instruction if we have already visited it before.
+    if (!Visited.insert(I).second)
+      continue;
 
     if (isa<PHINode>(I))
       continue;
@@ -234,9 +236,13 @@ BinaryOperator *
 RecurrenceInfo::digRecurrence(Instruction *V,
                               Instruction::BinaryOps BOWithConstOpToMatch) {
   SmallVector<Instruction *> Worklist;
+  SmallPtrSet<Instruction *, 16> Visited;
   Worklist.push_back(V);
   while (!Worklist.empty()) {
     Instruction *I = Worklist.pop_back_val();
+    // Skip this instruction if we have already visited it before.
+    if (!Visited.insert(I).second)
+      continue;
 
     // Don't add a PHI's operands to the Worklist.
     if (isa<PHINode>(I))
@@ -376,6 +382,71 @@ CRCTable HashRecognize::genSarwateTable(const APInt &GenPoly,
   return Table;
 }
 
+/// Perform polynomial (GF(2)) floor division. This is based on the
+/// floor_division(S, P) algorithm in
+/// https://www.corsix.org/content/barrett-reduction-polynomials. Note that the
+/// maximum degree of the returned polynomial is
+/// max(0, deg(Dividend) - deg(Divisor)), but the bit width will be the same as
+/// that of Dividend.
+static APInt floorDivideGF2(APInt Dividend, APInt Divisor) {
+  assert(!Divisor.isZero() && "Cannot divide by zero");
+
+  // Extend the divisor bit width to match the dividend.
+  Divisor = Divisor.zext(Dividend.getBitWidth());
+
+  // Note that getActiveBits returns deg+1, but the computation below
+  // still holds.
+  unsigned DivisorActiveBits = Divisor.getActiveBits();
+
+  // Q = 0
+  APInt Quotient = APInt::getZero(Dividend.getBitWidth());
+  // S != 0 and deg(S) >= deg(P)
+  // (S != 0 implied by DivisorActiveBits > 0)
+  while (Dividend.getActiveBits() >= DivisorActiveBits) {
+    // T = S[deg(S)] / P[deg(P)]
+    unsigned Shift = Dividend.getActiveBits() - DivisorActiveBits;
+    // Q = Q + T
+    Quotient.setBit(Shift);
+    // S = S - T * P
+    Dividend ^= Divisor.shl(Shift);
+  }
+  return Quotient;
+}
+
+/// Generate the constants for performing a Polynomial (GF(2)) Barrett Reduction
+/// according to Intel's Fast CRC Computation white paper with some adjustments
+/// to account for the fact that bit width and trip count can vary.
+std::pair<APInt, APInt>
+HashRecognize::genBarrettConstants(const PolynomialInfo &Info) {
+  unsigned BW = Info.RHS.getBitWidth();
+  unsigned TC = Info.TripCount;
+
+  // Recover the full generating polynomial in normal form by reflecting the LE
+  // case and adding the implied x^BW term.
+  // deg(P(x)) = BW due to the implied term, and thus P(x) must fit in exactly
+  // BW+1 bits.
+  APInt FullGenPoly =
+      (Info.IsBigEndian ? Info.RHS : Info.RHS.reverseBits()).zext(BW + 1);
+  FullGenPoly.setBit(BW);
+
+  // Calculate mu = floor(x^(BW+TC) / P(x)).
+  // deg(mu) <= deg(x^(BW+TC)) - deg(P(x)) = BW+TC - BW = TC, and thus mu must
+  // fit in at most TC+1 bits.
+  unsigned DivBW = BW + TC + 1;
+  APInt Mu = floorDivideGF2(APInt::getOneBitSet(DivBW, BW + TC), FullGenPoly)
+                 .trunc(TC + 1);
+
+  // In the bit-reflected (little-endian) case, mu and P(x) must be
+  // bit-reflected across their respective widths for the corresponding Barrett
+  // reduction steps.
+  if (!Info.IsBigEndian) {
+    Mu = Mu.reverseBits();
+    FullGenPoly = FullGenPoly.reverseBits();
+  }
+
+  return {Mu, FullGenPoly};
+}
+
 /// Checks that \p P1 and \p P2 are used together in an XOR in the use-def chain
 /// of \p SI's condition, ignoring any casts. The purpose of this function is to
 /// ensure that LHSAux from the SimpleRecurrence is used correctly in the CRC
@@ -393,6 +464,7 @@ CRCTable HashRecognize::genSarwateTable(const APInt &GenPoly,
 static bool isConditionalOnXorOfPHIs(const SelectInst *SI, const PHINode *P1,
                                      const PHINode *P2, const Loop &L) {
   SmallVector<const Instruction *> Worklist;
+  SmallPtrSet<const Instruction *, 16> Visited;
 
   // matchConditionalRecurrence has already ensured that the SelectInst's
   // condition is an Instruction.
@@ -400,6 +472,9 @@ static bool isConditionalOnXorOfPHIs(const SelectInst *SI, const PHINode *P1,
 
   while (!Worklist.empty()) {
     const Instruction *I = Worklist.pop_back_val();
+    // Skip this instruction if we have already visited it before.
+    if (!Visited.insert(I).second)
+      continue;
 
     // Don't add a PHI's operands to the Worklist.
     if (isa<PHINode>(I))
@@ -443,11 +518,12 @@ std::variant<PolynomialInfo, StringRef> HashRecognize::recognizeCRC() const {
   BasicBlock *Latch = L.getLoopLatch();
   BasicBlock *Exit = L.getExitBlock();
   const PHINode *IndVar = L.getCanonicalInductionVariable();
-  if (!Latch || !Exit || !IndVar || L.getNumBlocks() != 1)
+  if (!Latch || !Exit || !IndVar || L.getNumBlocks() != 1 ||
+      !L.getLatchCmpInst())
     return "Loop not in canonical form";
   unsigned TC = SE.getSmallConstantTripCount(&L);
-  if (!TC || TC % 8)
-    return "Unable to find a small constant byte-multiple trip count";
+  if (!TC)
+    return "Unable to find a small constant trip count";
 
   auto R = getRecurrences(Latch, IndVar, L);
   if (!R)
@@ -576,6 +652,13 @@ void HashRecognize::print(raw_ostream &OS) const {
   }
   OS.indent(2) << "Computed CRC lookup table:\n";
   genSarwateTable(Info.RHS, Info.IsBigEndian).print(OS);
+  OS.indent(2) << "Computed CRC Barrett constants:\n";
+  auto [Mu, FullGenPoly] = genBarrettConstants(Info);
+  OS << "Mu = ";
+  Mu.print(OS, false);
+  OS << ", FullGenPoly = ";
+  FullGenPoly.print(OS, false);
+  OS << "\n";
 }
 
 #if !defined(NDEBUG) || defined(LLVM_ENABLE_DUMP)