+++ llvm/docs/ReleaseNotes.md
@@ -6,8 +6,6 @@ ReleaseNotes.md and ReleaseNotesTemplate.txt. -->
 
 # LLVM {{env.config.release}} Release Notes
 
-```{contents}
-```
 
 ````{only} PreRelease
 ```{warning} These are in-progress notes for the upcoming LLVM {{env.config.release}}
@@ -60,6 +58,10 @@ Makes programs 10x faster by doing Special New Thing.
 
 ### Changes to TableGen
 
+* `!cond` operator short-circuits at the first `true` condition.  Subsequent
+  `condition : value` pairs, along with their corresponding side effects,
+  are left unresolved.
+
 ### Changes to Interprocedural Optimizations
 
 ### Changes to Vectorizers
@@ -67,6 +69,9 @@ Makes programs 10x faster by doing Special New Thing.
 ### Changes to the AArch64 Backend
 
 ### Changes to the AMDGPU Backend
+
+* Replaced `xnack` and `sramecc` target features with `amdgpu.xnack`
+  and `amdgpu.sramecc` module flags.
 
 ### Changes to the ARM Backend
 
@@ -103,6 +108,8 @@ Makes programs 10x faster by doing Special New Thing.
 ### Changes to the Debug Info
 
 ### Changes to the LLVM tools
+
+* llvm-mca no longer defaults -mcpu to "native"
 
 ### Changes to LLDB